Split out RegistryHelper Move IsElevated to Program
